Add notification preference controls
